Move HijackedResponse and CloseWriter to a separate file.
